Keep track of the LESS conversion in the repository
We are relying on LESS to manage the stylesheets as this is easier
to import code from DaPlayer. However, since LESS isn't part of the
default stack, we need to define our own converter but the latter
won't be used on GitHub pages.

Thus, we need to keep the generated CSS as a static file that will
be copied to the output directory at generation time.